requirements.txt
pytest
requests
pytest-html
pytest-xdist
pytest_ordering
pytest-rerunfailures
allure-pytest